requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
alluka==0.3.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorlog==6.8.2
deep-translator==1.11.4
deprecation==2.1.0
exceptiongroup==1.2.2
frozenlist==1.4.1
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
gotrue==2.9.1
h11==0.14.0
h2==4.1.0
hikari==2.1.0
hikari-lightbulb==2.3.5.post1
hikari-miru==4.1.1
hpack==4.0.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
multidict==6.1.0
packaging==24.1
postgrest==0.17.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
qrcode==8.0
realtime==2.0.5
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
storage3==0.8.0
StrEnum==0.4.15
supabase==2.8.1
supafunc==0.6.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
websockets==13.1
yarl==1.13.1
